Add database-generated geometry library text files.
They're created using the Protein Geometry Database, and the actual data is
downloaded from the PGD graph screen.

It is a set of 4 tab-delimited text files containing a list of
10-degree phi,psi bins. Each bin has an observation count and an
average and standard deviation for all backbone bond lengths and angles.

The data come from a search of all structures at <=1.0A, with
structures with >90% sequence identity removed. The search only
includes residues with B factor < 25 in that residue and the single
preceding and following residues. It's separated into 4 sets: Gly, Pro,
XPro, and the remaining 18 amino acids.