get_pdb_rmsd_stats.sh: Computes statistics from resolution slices.
Uses PDBSelect 90% threshold to grab 50 structures from a given
resolution range. 50 was chosen because that number fell within a 0.1 A
slice at all of the resolutions in PDBSelect from 1.0 to 3.0.

There are a few site-specific changes needed to run this. Set PDBSELECT
to the location of a downloaded PDBSelect file, and if necessary, change
the exported PYTHONPATH.