angles: Add an easy way to map a geometry name to the database's name.
Prototype:
average_name get_database_attribute_average_name(geometry_name)

This will accept a name in the form 'a1', 'l1', etc and return the name
that the database uses to reference the average ('a1Avg', 'L1Avg', etc).

Switch get_geom_rmsd.py to use this instead of its own code, because
Phenix also needs it so it was duplicated.