Allow passing multiple PDBs to get_geom_rmsd.py with multiple arguments, e.g. -s STRUCTURE1 -s STRUCTURE2.
Doesn't yet do within-group RMSD, just pairwise to the primary. If the 
compared residue is bad in any structure, ignore it altogether. The 
eventual goal is to do within-group per-residue RMSD as well as some 
sort of overall groupwise RMSD.

The groups are handled as a list of 'pdb_container' objects. This was 
necessary because filenames are invalid dictionary keys, and we use 
dictionaries to store a lot of the info. Since we're now in nested loops 
fairly often and Python lacks a 'break 2' syntax, use exceptions to 
break out of the nested loops.

Also fix up the import of conformation_dependent_geometry.angles so that 
angles.optlist is initialized properly and we don't error out.

My emacs complained about the encoding line (UTF-8) so I changed it to 
utf-8.